backend_python/requirements.txt
flask
flask-cors
gunicorn
web3
google-generativeai
